Move tests to Specification.Tests
